requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
Django==1.11.7
djangorestframework==3.7.1
kombu==4.1.0
olefile==0.44
Pillow==4.3.0
pytz==2017.3
redis==2.10.6
vine==1.1.4
wheel==0.24.0
